Move link provider register to bookmarkediting
